comments work now!

+++ comments/views.py
@@ -2,48 +2,66 @@ from django.shortcuts import render
 from .forms import CommentForm
 from .models import Comment
 from django.shortcuts import get_object_or_404, redirect
+from ideas.models import Idea
+import re
 
 def list_of_comments(request, idea_id, campaign_id):
     comments = Comment.objects.all().filter(idea_id=idea_id)
     context = {'comments': comments}
-    return render(request, 'comments/comments_list.html', context)
+    return render(request, 'ideas/idea_details.html', context)
 
-def comment_details(request, comment_id):
+def comment_details(request, comment_id, idea_id, campaign_id):
     comment = get_object_or_404(Comment, id=comment_id)
     
     if request.method == 'GET':
         form = CommentForm(instance=comment)
-        context =  {'form': form, 'comment': comment}
+        context =  {'form': form, 'comment': comment, 'idea_id': idea_id}
         return render(request, 'comments/comment_details.html', context)
     else:
         try:
-            form = CommentForm(request.POST, instance=commnet)
+            form = CommentForm(request.POST, instance=comment)
             form.save()
-            return redirect('comments:list_of_comments')
+            return redirect('comments:list_of_comments', idea_id, campaign_id)
         except ValueError:
             context = {'form': CommentForm(), 'error': 'Bad data try again'}
             return render(request, 'comments/comment_details.html', context)
     
 
-def comment_create(request):
+def comment_create(request, idea_id, campaign_id):
+    user = request.user
+    
     if request.method == 'GET':
-        return render(request, 'comments/comment_form.html', {'form': CommentForm()})
+        form = CommentForm(initial={'idea_id': idea_id})
+        return render(request, 'comments/comment_form.html', {'form': form})
     else:
         try:
-            form = CommentForm(request.POST)
+            form = CommentForm(request.POST, request.FILES, initial={'idea_id': idea_id})
             form.save()
-            return redirect('comments:list_of_comments')
+            return redirect('campaigns:ideas:comments:list_of_comments', idea_id, campaign_id)
         except ValueError:
             context = {'form': CommentForm(), 'error': 'Bad data try again'}
             return render(request, 'comments/comment_form.html', context)
         
-def comment_delete(request, comment_id):
+def comment_delete(request, comment_id, campaign_id):
+    user = request.user
     comment = get_object_or_404(Comment, id=comment_id)
-    comment.delete()
-    return redirect('comments:list_of_comments')
-
-def comment_update(request, comment_id):
+    if user.username == idea.author:
+        comment.delete()
+        return redirect('comments:list_of_comments', idea_id, campaign_id)
+    else:
+        return render(request, 'comments/permission_error.html')
+    
+def comment_update(request, comment_id, idea_id, campaign_id):
     comment = Comment.objects.get(id=comment_id)
-    form = CommentForm(instance=comment)
+    idea = Idea.objects.get(id=idea_id)
+    
+    if request.method == 'POST':
+        form = CommentForm(request.POST, instance=comment)
+        if form.is_valid():
+            form.save()
+            return redirect('comments:list_of_comments', idea_id, campaign_id)
+    else:
+        form = CommentForm(instance=comment)
+
     return render(request, 'comments/comment_update.html', {'form': form})
